Refactored

+++ src/main/java/com/example/java/test/junior/developer/controller/QuestionController.java
@@ -1,11 +1,18 @@
 package com.example.java.test.junior.developer.controller;
+
 import com.example.java.test.junior.developer.dto.QuestionDto;
 import com.example.java.test.junior.developer.service.QuestionService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.*;
-
-import javax.validation.Valid;
 import java.util.List;
+import javax.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
@@ -25,13 +32,13 @@ public class QuestionController {
     }
 
     @PutMapping("/{id}")
-    public QuestionDto updateQuestion(@PathVariable int id,
+    public QuestionDto updateQuestion(@PathVariable Long id,
                                       @RequestBody @Valid QuestionDto question){
         return questionService.updateQuestion(id, question);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteQuestion(@PathVariable int id) {
+    public void deleteQuestion(@PathVariable Long id) {
         questionService.deleteQuestion(id);
     }
 }
