+++ src/main/java/com/example/java/test/junior/developer/dto/CategoryDto.java
@@ -1,8 +1,6 @@
 package com.example.java.test.junior.developer.dto;
 
 import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.NoArgsConstructor;
@@ -14,7 +12,7 @@ import lombok.Value;
 @AllArgsConstructor
 public class CategoryDto {
 
-    int id;
+    Long id;
 
     @NotBlank String name;
