+++ src/main/java/com/example/java/test/junior/developer/repository/CategoryRepository.java
@@ -3,5 +3,5 @@ package com.example.java.test.junior.developer.repository;
 import com.example.java.test.junior.developer.model.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryRepository extends JpaRepository<Category,Integer> {
+public interface CategoryRepository extends JpaRepository<Category, Long> {
 }
